refactor: 트랜잭션 관리 계층 추상화
