+++ app/src/main/java/com/techcourse/service/TxUserService.java
@@ -1,0 +1,54 @@
+package com.techcourse.service;
+
+import com.interface21.dao.DataAccessException;
+import com.interface21.jdbc.datasource.DataSourceUtils;
+import com.interface21.transaction.TransactionRollbackException;
+import com.techcourse.config.DataSourceConfig;
+import com.techcourse.domain.User;
+import java.sql.Connection;
+import java.sql.SQLException;
+import javax.sql.DataSource;
+
+public class TxUserService implements UserService {
+
+    private final UserService userService;
+
+    public TxUserService(final UserService userService) {
+        this.userService = userService;
+    }
+
+    @Override
+    public User findById(final long id) {
+        return userService.findById(id);
+    }
+
+    @Override
+    public void insert(final User user) {
+        userService.insert(user);
+    }
+
+    @Override
+    public void changePassword(final long id, final String newPassword, final String createBy) {
+        final DataSource dataSource = DataSourceConfig.getInstance();
+        final Connection connection = DataSourceUtils.getConnection(dataSource);
+
+        try {
+            connection.setAutoCommit(false);
+            userService.changePassword(id, newPassword, createBy);
+            connection.commit();
+        } catch (SQLException e) {
+            rollbackTransaction(connection);
+            throw new DataAccessException(e);
+        } finally {
+            DataSourceUtils.releaseConnection(connection, dataSource);
+        }
+    }
+
+    private void rollbackTransaction(final Connection connection) {
+        try {
+            connection.rollback();
+        } catch (SQLException e) {
+            throw new TransactionRollbackException("Transaction rollback failed", e);
+        }
+    }
+}
