+++ app/src/main/java/com/techcourse/service/UserService.java
@@ -1,63 +1,12 @@
 package com.techcourse.service;
 
-import com.interface21.dao.DataAccessException;
-import com.interface21.transaction.TransactionRollbackException;
-import com.techcourse.config.DataSourceConfig;
-import com.techcourse.dao.UserDao;
-import com.techcourse.dao.UserHistoryDao;
 import com.techcourse.domain.User;
-import com.techcourse.domain.UserHistory;
-import java.sql.Connection;
-import java.sql.SQLException;
-import javax.sql.DataSource;
 
-public class UserService {
+public interface UserService {
 
-    private final UserDao userDao;
-    private final UserHistoryDao userHistoryDao;
+    User findById(final long id);
 
-    public UserService(final UserDao userDao, final UserHistoryDao userHistoryDao) {
-        this.userDao = userDao;
-        this.userHistoryDao = userHistoryDao;
-    }
+    void insert(final User user);
 
-    public User findById(final long id) {
-        return userDao.findById(id);
-    }
-
-    public void insert(final User user) {
-        userDao.insert(user);
-    }
-
-    public void changePassword(final long id, final String newPassword, final String createBy) {
-        final Connection connection = getConnection(DataSourceConfig.getInstance());
-        final var user = findById(id);
-
-        try (connection) {
-            connection.setAutoCommit(false);
-            user.changePassword(newPassword);
-            userDao.update(connection, user);
-            userHistoryDao.log(connection, new UserHistory(user, createBy));
-            connection.commit();
-        } catch (SQLException e) {
-            rollbackTransaction(connection);
-            throw new DataAccessException(e);
-        }
-    }
-
-    private void rollbackTransaction(final Connection connection) {
-        try {
-            connection.rollback();
-        } catch (SQLException e) {
-            throw new TransactionRollbackException("Transaction rollback failed", e);
-        }
-    }
-
-    private Connection getConnection(final DataSource dataSource) {
-        try {
-            return dataSource.getConnection();
-        } catch (SQLException e) {
-            throw new DataAccessException(e);
-        }
-    }
+    void changePassword(final long id, final String newPassword, final String createBy);
 }
